+++ ambry-account/src/main/java/com/github/ambry/account/AccountInfoMap.java
@@ -172,14 +172,6 @@ class AccountInfoMap {
   }
 
   /**
-   * Gets all the {@link Account}s' Id in this {@code AccountInfoMap} in a {@link Collection}.
-   * @return A {@link Collection} of all the {@link Account}s' Id in this map.
-   */
-  Collection<Short> getAccountIds() {
-    return Collections.unmodifiableCollection(idToAccountMap.keySet());
-  }
-
-  /**
    * Return true if there is no accounts in this info map.
    * @return True when there is no accounts.
    */
