+++ ambry-account/src/test/java/com/github/ambry/account/JsonAccountServiceTest.java
@@ -69,7 +69,7 @@ public class JsonAccountServiceTest {
 
     accountJsonFilePath = configPath.resolve("accounts.json");
 
-    accountService = new JsonAccountService(accountJsonFilePath, new AccountServiceMetrics(new MetricRegistry()), null,
+    accountService = new JsonAccountService(accountJsonFilePath, new AccountServiceMetrics(new MetricRegistry(), false), null,
         accountConfig);
     accountService.init();
   }
