+++ ambry-account/src/test/java/com/github/ambry/account/MockHelixAccountServiceFactory.java
@@ -46,7 +46,7 @@ public class MockHelixAccountServiceFactory extends HelixAccountServiceFactory {
   public MockHelixAccountServiceFactory(VerifiableProperties verifiableProperties, MetricRegistry metricRegistry,
       Notifier<String> notifier, String updaterThreadPrefix, Router router) {
     super(new HelixPropertyStoreConfig(verifiableProperties), new HelixAccountServiceConfig(verifiableProperties),
-        new AccountServiceMetrics(metricRegistry), notifier);
+        new AccountServiceMetrics(metricRegistry, false), notifier);
     HelixAccountServiceConfig config = new HelixAccountServiceConfig(verifiableProperties);
     this.updaterThreadPrefix = updaterThreadPrefix;
     this.router = router;
